Refactor non-vscode funcs to ease testing
