+++ src/papyrus-lang-vscode/src/PapyrusGame.ts
@@ -1,20 +1,8 @@
-import * as fs from 'fs';
-import { promisify } from 'util';
-
-import { xml2js } from 'xml-js';
-
-import { workspace, Uri, RelativePattern } from 'vscode';
-
-import { PyroGameToPapyrusGame } from './features/PyroTaskDefinition';
-
-const readFile = promisify(fs.readFile);
-
 export enum PapyrusGame {
     fallout4 = 'fallout4',
     skyrim = 'skyrim',
     skyrimSpecialEdition = 'skyrimSpecialEdition',
 }
-
 const displayNames = new Map([
     [PapyrusGame.fallout4, 'Fallout 4'],
     [PapyrusGame.skyrim, 'Skyrim'],
@@ -57,40 +45,38 @@ export function getGames(): PapyrusGame[] {
     return (Object.keys(PapyrusGame) as (keyof typeof PapyrusGame)[]).map((k) => PapyrusGame[k]);
 }
 
-export async function getWorkspaceGameFromProjects(ppjFiles: Uri[]): Promise<PapyrusGame | undefined> {
-    let game: string | undefined = undefined;
-    if (!ppjFiles) {
-        return undefined;
+export function getRegistryKeyForGame(game: PapyrusGame) {
+    switch (game) {
+        case PapyrusGame.fallout4:
+            return 'Fallout4';
+        case PapyrusGame.skyrim:
+            return 'Skyrim';
+        case PapyrusGame.skyrimSpecialEdition:
+            return 'Skyrim Special Edition';
     }
-
-    for (const ppjFile of ppjFiles) {
-        game = await getWorkspaceGameFromProjectFile(ppjFile.fsPath);
-        if (game) {
-            break;
-        }
-    }
-
-    if (!game || !PyroGameToPapyrusGame[game as keyof typeof PyroGameToPapyrusGame]) {
-        return undefined;
-    }
-
-    return PyroGameToPapyrusGame[game as keyof typeof PyroGameToPapyrusGame] as unknown as PapyrusGame;
 }
 
-export async function getWorkspaceGameFromProjectFile(projectFile: string): Promise<PapyrusGame | undefined> {
-    const xml = await readFile(projectFile, { encoding: 'utf-8' });
-    // TODO: Annoying type cast here:
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const results = xml2js(xml, { compact: true, trim: true }) as Record<string, any>;
-
-    return results['PapyrusProject']['_attributes']['Game'];
+export function getDevelopmentCompilerFolderForGame(game: PapyrusGame) {
+    switch (game) {
+        case PapyrusGame.fallout4:
+            return 'fallout4';
+        case PapyrusGame.skyrim:
+            return 'does-not-exist';
+        case PapyrusGame.skyrimSpecialEdition:
+            return 'skyrim';
+    }
 }
 
-export async function getWorkspaceGame(): Promise<PapyrusGame | undefined> {
-    if (!workspace.workspaceFolders) {
-        return undefined;
-    }
+export function getDefaultFlagsFileNameForGame(game: PapyrusGame) {
+    return game === PapyrusGame.fallout4 ? 'Institute_Papyrus_Flags.flg' : 'TESV_Papyrus_Flags.flg';
+}
 
-    const ppjFiles: Uri[] = await workspace.findFiles(new RelativePattern(workspace.workspaceFolders[0], '**/*.ppj'));
-    return getWorkspaceGameFromProjects(ppjFiles);
+const executableNames = new Map([
+    [PapyrusGame.skyrim, 'Skyrim.exe'],
+    [PapyrusGame.fallout4, 'Fallout4.exe'],
+    [PapyrusGame.skyrimSpecialEdition, 'SkyrimSE.exe'],
+]);
+
+export function getExecutableNameForGame(game: PapyrusGame) {
+    return executableNames.get(game)!;
 }
