+++ src/papyrus-lang-vscode/src/features/commands/AttachDebuggerCommand.ts
@@ -1,7 +1,7 @@
 import { injectable } from 'inversify';
 import { debug } from 'vscode';
 import { IPapyrusDebugConfiguration } from '../../debugger/PapyrusDebugSession';
-import { PapyrusGame, getShortDisplayNameForGame } from '../../PapyrusGame';
+import { PapyrusGame, getShortDisplayNameForGame } from "../../PapyrusGame";
 import { GameCommandBase } from './GameCommandBase';
 
 @injectable()
