+++ src/papyrus-lang-vscode/src/features/commands/GameCommandBase.ts
@@ -1,6 +1,6 @@
 import { injectable, unmanaged } from 'inversify';
 import { commands, Disposable } from 'vscode';
-import { PapyrusGame, getGames } from '../../PapyrusGame';
+import { PapyrusGame, getGames } from "../../PapyrusGame";
 
 @injectable()
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
